Fix a couple of tests on Win64

+++ root/meta/runInterpClassDef.C
@@ -12,7 +12,7 @@ struct Base {
 };
 struct Derived: public Base {
    // Make sure that this Base::IsA() is *not* virtual!
-   TClass *IsA() const { return (TClass*) (long) -1; }
+   TClass *IsA() const { return (TClass*) (intptr_t) -1; }
 };
 // Test for the call of Error function
 // from within the Inner struct (ROOT-7441)
@@ -28,7 +28,7 @@ int runInterpClassDef() {
 
   Derived d;
   Base* b = &d;
-  if (((long)b->IsA()) == -1)
+  if (((intptr_t)b->IsA()) == -1)
     return 1; // FAILURE
   return 0;
 }
